% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoarea-package.R, R/geoarea.R
\docType{package}
\name{geoarea}
\alias{geoarea}
\title{geoarea.}
\usage{
geoarea(x)
}
\arguments{
\item{x}{Rectangular object (matrix, \code{data.frame}, \pkg{tibble},
whatever) containing longitude and latitude coordinates of a polygon.}
}
\value{
A vector of two numbers quantifying the area and perimter of the
polygon 'x'.
}
\description{
Calculation is based on the "geographiclib" code of Charles Karney
<doi:10.1007/s00190-012-0578-z>, available at
\url{https://github.com/geographiclib/geographiclib-c}. These calculations
are available in many other libraries and R packages, and are notably
included in the 'PROJ' library, and used by many packages which depend on
that library, including \pkg{sf}. The \pkg{geosphere} package also includes
the C++ version of this code, but depends on the obsolete \pkg{sp} package.
}
\details{
In contrast to all other implementations, this function performs no
pre-processing or input checking whatsoever, and simply aims to provide the
fastest way to calculate areas of polygons expressed in longitudes and
latitudes.
}
\examples{
# Perimeter of Antarctica as documented in original 'geographiclib' code
lats <- c (-72.9, -71.9, -74.9, -74.3, -77.5, -77.4, -71.7, -65.9, -65.7,
           -66.6, -66.9, -69.8, -70.0, -71.0, -77.3, -77.9, -74.7)
lons <- c (-74, -102, -102, -131, -163, 163, 172, 140, 113, 88, 59, 25, -4,
           -14, -33, -46, -61)
xy <- cbind (lons, lats)
geoarea (xy)
}
